secret/python/secret/pyproject.toml
[project]
name = "secret"
version = "0.1.0"
description = "A simple secrets manager based on GNUPG."
authors = [ { name="Steve Hay", email="me@stevenhay.com" } ]
requires-python = ">=3.7"
dependencies = [
    "python-gnupg>=0.5.4"
]

[project.license]
text = "MIT"

[project.optional-dependencies]
dev = [
    "ruff",
]

[project.scripts]
secret = "secret:main"

[build-system]
requires = ["setuptools >= 39.2"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["secret"]
